+++ tests/testthat/test-roxytest-testexamples-02-pipeline.R
@@ -2,7 +2,7 @@
 
 # File R/"02-pipeline.R": @testexamples
 
-test_that("[unknown alias] @ L117", {
+test_that("[unknown alias] @ L113", {
   
   library(bregr)
   # 1. Pipeline -------------------------